# "Breadcrumb" room sorting algorithm