… src/i18n/strings/en_EN.json