**Caution**: This does not use a labs flag and alters how the room list sorting is done.

*This is being opened now to give people a point to test off of without impacting develop*.

Fixes https://github.com/vector-im/riot-web/issues/8531
Fixes https://github.com/vector-im/riot-web/issues/8122
Fixes https://github.com/vector-im/riot-web/issues/7826
Fixes https://github.com/vector-im/riot-web/issues/7803 ~~(well, not yet...)~~
Fixes https://github.com/vector-im/riot-web/issues/8774

Known bugs:
* [x] Reading rooms on other devices does not cause them to fall down
* [x] ~~UNCONFIRMED:~~ Potentially high CPU usage trying to math all this out
* [x] Direct chats go nuts and end up in the wrong order after refreshing

Room for improvement:
* Tags could be magically calculated rather than bluntly throwing away our incremental lists
* Same applies to direct chat tagging