@turt2live 2019-02-14
Early performance characteristics (based on my account; before optimization):
* Initial room list generation: 0.13ms -> 394ms
* Incremental room list generation: 0.26ms -> 1.5ms

I'm not entirely sure how the old algorithm achieves 0.13ms initial generation time, but it is at least consistent given the same algorithm is used for incremental changes. Suspicion for the new initial list generation taking 400ms is lack of caching, so we'll see how that goes.