@turt2live 2019-02-19
sorry, this technically wasn't back up for review while I'm investigating why its so bad at performance. Intention is to make it transparent, which will probably happen by the time we cut another point release.