@turt2live 2019-02-19
I lied: I thought this would take longer to optimize, but I was wrong. After staring at profiles for a few hours and making small changes I've managed to cut it down to the following characteristics:

* First sort time: 2ms worst case, under 1ms most cases.
* Incremental sort time: 1ms worst case, under 0.4ms most cases